refactor: tighten multisig public surface
