+++ src/server/Handler.test-d.ts
@@ -6,7 +6,6 @@ import { describe, expectTypeOf, test } from 'vp/test'
 
 import * as CliAuth from './CliAuth.js'
 import * as Handler from './Handler.js'
-import * as Multisig from './Multisig.js'
 
 describe('codeAuth options', () => {
   test('supports chain-agnostic chains/transports configuration', () => {
@@ -38,20 +37,6 @@ describe('codeAuth options', () => {
         return request.headers.get('x-forwarded-for') ?? 'unknown'
       },
     })
-  })
-})
-
-describe('relay options', () => {
-  test('accepts public multisig store options', () => {
-    void Handler.relay({
-      multisig: {
-        claimTtl: 30_000,
-        store: Multisig.memoryStore(),
-      },
-    })
-    expectTypeOf<Multisig.Store>().toMatchTypeOf<
-      NonNullable<Handler.relay.Options['multisig']>['store']
-    >()
   })
 })
