+++ src/server/Multisig.test.ts
@@ -7,6 +7,7 @@ import { afterAll, beforeAll, beforeEach, describe, expect, test } from 'vp/test
 
 import { createServer, type Server } from '../../test/utils.js'
 import * as Handler from './Handler.js'
+import * as MultisigInternal from './internal/handlers/multisig.js'
 import * as Multisig from './Multisig.js'
 
 const privateKey_1 = '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80'
@@ -305,12 +306,6 @@ describe('Handler.relay multisig', () => {
 })
 
 describe('handleRawTransaction', () => {
-  test('behavior: returns null status for unknown operations', async () => {
-    await expect(
-      Multisig.getStatus({ id: `0x${'ff'.repeat(32)}`, store: Multisig.memoryStore() }),
-    ).resolves.toMatchInlineSnapshot(`null`)
-  })
-
   test('behavior: rejects first approvals when no config can be resolved', async () => {
     const owner = Account.fromSecp256k1(privateKey_1)
     const account = Account.fromMultisig({
@@ -332,7 +327,7 @@ describe('handleRawTransaction', () => {
     )
 
     await expect(
-      Multisig.handleRawTransaction({
+      MultisigInternal.handleRawTransaction({
         getClient: (() => undefined) as never,
         method: 'eth_sendRawTransaction',
         request: { params: [serialized] },
@@ -365,7 +360,7 @@ describe('handleRawTransaction', () => {
     } as never)
     let broadcastMethod: string | undefined
 
-    const hash = await Multisig.handleRawTransaction({
+    const hash = await MultisigInternal.handleRawTransaction({
       finalize: 'submitted',
       getClient: (() => ({
         request: async ({ method, params }: { method: string; params: unknown }) => {
@@ -418,13 +413,13 @@ describe('config resolution', () => {
       await account.signTransaction({ ...transaction, signatures: [signature_2] } as never),
     )
 
-    const id = (await Multisig.handleRawTransaction({
+    const id = (await MultisigInternal.handleRawTransaction({
       getClient: (() => undefined) as never,
       method: 'eth_sendRawTransaction',
       request: { params: [first] },
       store,
     })) as `0x${string}`
-    const hash = await Multisig.handleRawTransaction({
+    const hash = await MultisigInternal.handleRawTransaction({
       getClient: (() => ({
         request: async ({ params }: { params: unknown }) => hashRawTransaction(params),
       })) as never,
@@ -466,7 +461,7 @@ describe('config resolution', () => {
     } as never)
 
     await expect(
-      Multisig.handleRawTransaction({
+      MultisigInternal.handleRawTransaction({
         getClient: (() => undefined) as never,
         method: 'eth_sendRawTransaction',
         request: { params: [serialized] },
@@ -510,7 +505,7 @@ describe('config resolution', () => {
     )
     let submitted: `0x${string}` | undefined
 
-    const hash = await Multisig.handleRawTransaction({
+    const hash = await MultisigInternal.handleRawTransaction({
       getClient: (() => ({
         request: async ({ params }: { params: unknown }) => {
           submitted = Array.isArray(params) ? (params[0] as `0x${string}`) : undefined
